kasper: Adds Kasper support
The kernel accepts syscalls between 600 and 1200 and executes them
as the syscall - 600. We use this to detect syscall injected by
syzkaller testcases and to ignore all other syscalls.

kasan print_error_description also includes the actual address
instead of only the symbolized func_offset.

mm/kspecem/ includes a stub of weak definition if KSPECEM is enabled
to throw an error & empty defintion to ignore inserted hooks if KSPECEM
is disabled.

add asan-recover=0 param to fix IP reported by KASan with KDFSan
enabled:

This is a workaround to fix the instruction pointer reported in KASan
reports when KDFSan is enabled. Prior to this, the reported IPs would be
in KASan handlers (e.g., __asan_load4_noabort), because the handler
would be called by a KDFSan wrapper (e.g., dfsw$__asan_load4), so
_RET_IP_ would give the IP for that rather than the code with the bad
access. With 'asan-recover=0', the "non-noabort" handlers are called,
e.g., __asan_load4 instead of __asan_load4_noabort. Since these handlers
aren't aliases, KDFSan handles them correctly (i.e., there are no
wrappers being invoked), so the _RET_IP_ is that of the code with
the bad access.